+++ lib/dimigoin_flutter_plugin.dart
@@ -4,7 +4,8 @@ import 'dart:convert';
 import 'package:dio/dio.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 
-part 'services/dimigoin_login.dart';
+part 'services/dimigoin_account.dart';
+part 'models/dimigoin_user.dart';
 part 'services/dimigoin_meal.dart';
 part 'services/dalgeurak_service.dart';
 
@@ -15,14 +16,17 @@ final _storage = const FlutterSecureStorage();
 const apiUrl = "https://api.dimigo.in";
 
 
-DimigoinLogin _dimigoinLogin = DimigoinLogin();
+DimigoinAccount _dimigoinLogin = DimigoinAccount();
 
 late String _accessToken;
+late DimigoinUser _currentUser;
 
 class DimigoinFlutterPlugin {
   initializeApp() async {
     if (await _dimigoinLogin.checkNowLogin()) {
       _accessToken = await _dimigoinLogin.loadSavedToken();
+
+      await _dimigoinLogin.fetchAccountData();
     }
   }
 }