+++ lib/services/dimigoin_account.dart
@@ -1,11 +1,12 @@
 part of dimigoin_flutter_plugin;
 
 /// 디미고인 로그인 클래스
-class DimigoinLogin {
+class DimigoinAccount {
+
+  DimigoinUser get currentUser => _currentUser;
 
   /// 디미고인 계정에 로그인을 진행하는 함수입니다.
   /// OAuth 방식을 사용하여, 로그인에 성공할 경우 반환되는 AccessToken과 RefreshToken을 Secure Storage에 저장합니다.
-  /// 또한 디미고인 계정 정보를 불러와 Secure Storage에 저장합니다.
   ///
   /// @param [username] 사용자의 디미고인 계정 아이디 string형 변수입니다.
   /// @param [password] 사용자의 디미고인 계정 비밀번호 string형 변수입니다.
@@ -18,17 +19,12 @@ class DimigoinLogin {
         data: {"username": userName, "password": password},
       );
 
-      Response infoResponse = await _dio.get(
-        "$apiUrl/user/me",
-        options: Options(contentType: "application/json", headers: {'Authorization': 'Bearer ${authResponse.data['accessToken']}'}),
-      );
-
 
       _accessToken = authResponse.data['accessToken'];
-
       await _storage.write(key: "dimigoinAccount_accessToken", value: authResponse.data['accessToken']);
       await _storage.write(key: "dimigoinAccount_refreshToken", value: authResponse.data['refreshToken']);
-      await _storage.write(key: "dimigoinAccount_userInfo", value: json.encode(infoResponse.data['identity']));
+      await storeUserData();
+
       return true;
     } catch (e) {
       return false;
@@ -56,7 +52,7 @@ class DimigoinLogin {
   validateAccessToken() async {
     String? accessToken = await _storage.read(key: "dimigoinAccount_accessToken");
     try {
-      Response response = await _dio.get(
+      await _dio.get(
         "$apiUrl/user/me",
         options: Options(contentType: "application/json", headers: {'Authorization': 'Bearer $accessToken'}),
       );
@@ -101,8 +97,31 @@ class DimigoinLogin {
   /// @returns 저장되어있는 AccessToken을 반환합니다.
   loadSavedToken() async => await _storage.read(key: "dimigoinAccount_accessToken");
 
-  /// 현재 서비스에 로그인되어 있는 경우, 저장되어있는 사용자 정보를 불러오는 함수입니다.
+  /// 디미고인 서버에 저장되어있는 계정의 정보를 불러와 로컬 Storage에 저장합니다.
   ///
-  /// @returns 저장되어있는 사용자 정보를 반환합니다.
-  loadUserInfo() async => json.decode((await _storage.read(key: "dimigoinAccount_userInfo"))!);
+  /// @returns 저장에 성공할 경우 true, 실패할 경우 false를 반환합니다.
+  storeUserData() async {
+    try {
+      Response infoResponse = await _dio.get(
+        "$apiUrl/user/me",
+        options: Options(contentType: "application/json",
+            headers: {'Authorization': 'Bearer $_accessToken'}),
+      );
+
+      await _storage.write(key: "dimigoinAccount_userInfo", value: json.encode(infoResponse.data['identity']));
+      _currentUser = DimigoinUser.fromJson(infoResponse.data['identity']);
+
+      return true;
+    } catch (e) {
+      return false;
+    }
+  }
+
+  /// 현재 서비스에 로그인되어 있는 경우, 디미고인 API에서 받아온 최신 데이터로 패치를 진행하는 함수입니다.
+  fetchAccountData() async {
+    if (!(await validateAccessToken())) { await refreshAccessToken(); }
+
+    bool isSuccessStoreData = await storeUserData();
+    if (!isSuccessStoreData) { _currentUser = DimigoinUser.fromJson(json.decode((await _storage.read(key: "dimigoinAccount_userInfo"))!)); }
+  }
 }